# Proxy stats 7x performance improvement