--- a/Hermes/appscale/hermes/producers/proxy_stats.py
+++ b/Hermes/appscale/hermes/producers/proxy_stats.py
@@ -1,4 +1,4 @@
-import StringIO
+import cStringIO
 import csv
 import logging
 import socket
@@ -27,8 +27,117 @@ class BoundIpPortNotFound(Exception):
   pass
 
 
+class HAProxyStatsParsingError(Exception):
+  pass
+
+
+HEADER_1_5_PLUS = [
+  "pxname", "svname", "qcur", "qmax", "scur", "smax", "slim", "stot", "bin",
+  "bout", "dreq", "dresp", "ereq", "econ", "eresp", "wretr", "wredis", "status",
+  "weight", "act", "bck", "chkfail", "chkdown", "lastchg", "downtime", "qlimit",
+  "pid", "iid", "sid", "throttle", "lbtot", "tracked", "type", "rate",
+  "rate_lim", "rate_max", "check_status", "check_code", "check_duration",
+  "hrsp_1xx", "hrsp_2xx", "hrsp_3xx", "hrsp_4xx", "hrsp_5xx", "hrsp_other",
+  "hanafail", "req_rate", "req_rate_max", "req_tot", "cli_abrt", "srv_abrt",
+  "comp_in", "comp_out", "comp_byp", "comp_rsp", "lastsess", "last_chk",
+  "last_agt", "qtime", "ctime", "rtime", "ttime"
+]
+ACCEPTABLE_HEADER = [
+  "pxname", "svname", "qcur", "qmax", "scur", "smax", "slim", "stot", "bin",
+  "bout", "dreq", "dresp", "ereq", "econ", "eresp", "wretr", "wredis", "status",
+  "weight", "act", "bck", "chkfail", "chkdown", "lastchg", "downtime", "qlimit",
+  "pid", "iid", "sid", "throttle", "lbtot", "tracked", "type", "rate",
+  "rate_lim", "rate_max", "check_status", "check_code", "check_duration",
+  "hrsp_1xx", "hrsp_2xx", "hrsp_3xx", "hrsp_4xx", "hrsp_5xx", "hrsp_other",
+  "hanafail", "req_rate", "req_rate_max", "req_tot", "cli_abrt", "srv_abrt"
+]
+
+# HAProxy stats fields numeration.
+# For more details see
+# https://cbonte.github.io/haproxy-dconv/1.5/configuration.html#9.1
+PXNAME = 0
+SVNAME = 1
+QCUR = 2
+QMAX = 3
+SCUR = 4
+SMAX = 5
+SLIM = 6
+STOT = 7
+BIN = 8
+BOUT = 9
+DREQ = 10
+DRESP = 11
+EREQ = 12
+ECON = 13
+ERESP = 14
+WRETR = 15
+WREDIS = 16
+STATUS = 17
+WEIGHT = 18
+ACT = 19
+BCK = 20
+CHKFAIL = 21
+CHKDOWN = 22
+LASTCHG = 23
+DOWNTIME = 24
+QLIMIT = 25
+PID = 26
+IID = 27
+SID = 28
+THROTTLE = 29
+LBTOT = 30
+TRACKED = 31
+TYPE = 32
+RATE = 33
+RATE_LIM = 34
+RATE_MAX = 35
+CHECK_STATUS = 36
+CHECK_CODE = 37
+CHECK_DURATION = 38
+HRSP_1XX = 39
+HRSP_2XX = 40
+HRSP_3XX = 41
+HRSP_4XX = 42
+HRSP_5XX = 43
+HRSP_OTHER = 44
+HANAFAIL = 45
+REQ_RATE = 46
+REQ_RATE_MAX = 47
+REQ_TOT = 48
+CLI_ABRT = 49
+SRV_ABRT = 50
+COMP_IN = 51
+COMP_OUT = 52
+COMP_BYP = 53
+COMP_RSP = 54
+LASTSESS = 55
+LAST_CHK = 56
+LAST_AGT = 57
+QTIME = 58
+CTIME = 59
+RTIME = 60
+TTIME = 61
+
+COLUMNS_NUMBER = 62
+
+HAPROXY_STATS_STR_COLUMNS = [
+  PXNAME, SVNAME, STATUS, CHECK_STATUS, LAST_CHK, LAST_AGT
+]
+HAPROXY_STATS_INT_COLUMNS = [
+  # All stats columns but:
+  # pxname, svname, status, check_status, last_chk, last_agt
+  QCUR, QMAX, SCUR, SMAX, SLIM, STOT, BIN, BOUT, DREQ, DRESP,
+  EREQ, ECON, ERESP, WRETR, WREDIS, WEIGHT, ACT, BCK, CHKFAIL, CHKDOWN,
+  LASTCHG, DOWNTIME, QLIMIT, PID, IID, SID, THROTTLE, LBTOT, TRACKED, TYPE,
+  RATE, RATE_LIM, RATE_MAX, CHECK_CODE, CHECK_DURATION, HRSP_1XX,
+  HRSP_2XX, HRSP_3XX, HRSP_4XX, HRSP_5XX, HRSP_OTHER, HANAFAIL, REQ_RATE,
+  REQ_RATE_MAX, REQ_TOT, CLI_ABRT, SRV_ABRT, COMP_IN, COMP_OUT, COMP_BYP,
+  COMP_RSP, LASTSESS, QTIME, CTIME, RTIME, TTIME
+]
+
+
 @include_list_name('proxy.listener')
-@attr.s(cmp=False, hash=False, slots=True, frozen=True)
+@attr.s(hash=False, slots=True, frozen=True)
 class HAProxyListenerStats(object):
   """
   For more details see
@@ -53,7 +162,7 @@ class HAProxyListenerStats(object):
 
 
 @include_list_name('proxy.frontend')
-@attr.s(cmp=False, hash=False, slots=True, frozen=True)
+@attr.s(hash=False, slots=True, frozen=True)
 class HAProxyFrontendStats(object):
   """
   For more details see
@@ -93,7 +202,7 @@ class HAProxyFrontendStats(object):
 
 
 @include_list_name('proxy.backend')
-@attr.s(cmp=False, hash=False, slots=True, frozen=True)
+@attr.s(hash=False, slots=True, frozen=True)
 class HAProxyBackendStats(object):
   """
   For more details see
@@ -148,7 +257,7 @@ class HAProxyBackendStats(object):
 
 
 @include_list_name('proxy.server')
-@attr.s(cmp=False, hash=False, slots=True, frozen=True)
+@attr.s(hash=False, slots=True, frozen=True)
 class HAProxyServerStats(object):
   """
   For more details see
@@ -210,26 +319,12 @@ class HAProxyServerStats(object):
   ttime = attr.ib()  # the avg total session time in ms over the 1024 last
 
 
-ALL_HAPROXY_FIELDS = set(
-  attr.fields_dict(HAProxyListenerStats).keys() +
-  attr.fields_dict(HAProxyFrontendStats).keys() +
-  attr.fields_dict(HAProxyBackendStats).keys() +
-  attr.fields_dict(HAProxyServerStats).keys()
-) - {'private_ip', 'port'}    # HAProxy stats doesn't include IP/Port columns
-                              # But we add these values by ourselves
-
-KNOWN_NON_INTEGER_FIELDS = {
-  'pxname', 'svname', 'status', 'check_status', 'last_chk', 'last_agt'
-}
-INTEGER_FIELDS = set(ALL_HAPROXY_FIELDS) - KNOWN_NON_INTEGER_FIELDS
-
-
 class InvalidHAProxyStats(ValueError):
   pass
 
 
 @include_list_name('proxy')
-@attr.s(cmp=False, hash=False, slots=True, frozen=True)
+@attr.s(hash=False, slots=True, frozen=True)
 class ProxyStats(object):
   """
   Object of ProxyStats is kind of structured container for all haproxy stats
@@ -249,32 +344,20 @@ class ProxyStats(object):
   listeners_count = attr.ib()  # number of listeners serving this proxy
 
 
-@attr.s(cmp=False, hash=False, slots=True, frozen=True)
+@attr.s(hash=False, slots=True, frozen=True)
 class ProxiesStatsSnapshot(object):
   utc_timestamp = attr.ib()  # UTC timestamp
   proxies_stats = attr.ib(metadata={Meta.ENTITY_LIST: ProxyStats})
-
-
-def _get_field_value(row, field_name):
-  """ Private method for getting value from csv cell. """
-  if field_name not in row:
-    return MISSED
-  value = row[field_name]
-  if not value:
-    return None
-  if field_name in INTEGER_FIELDS:
-    return int(value)
-  return value
 
 
 def get_stats(socket_path):
   client = socket.socket(socket.AF_UNIX, socket.SOCK_STREAM)
   client.connect(socket_path)
   try:
-    stats_output = StringIO.StringIO()
-    client.send('show stat\n')
+    stats_output = cStringIO.StringIO()
+    client.send(b'show stat\n')
     while True:
-      data = client.recv(1024)
+      data = client.recv(2048)
       if not data:
         break
       stats_output.write(data)
@@ -296,51 +379,227 @@ def get_frontend_ip_port(configs_dir, proxy_name):
                             .format(proxy_name, configs_dir))
 
 
-def get_connections(ip, port):
-  return sum(1 for conn in psutil.net_connections()
-             if conn.laddr == (ip, port) and conn.status == 'ESTABLISHED')
+def _convert_ints_and_missing(row):
+  if row[-1] == '':
+    row[-1] = MISSED
+  if len(row) < COLUMNS_NUMBER:
+    row += [MISSED] * (COLUMNS_NUMBER - len(row))
+  for index in HAPROXY_STATS_STR_COLUMNS:
+    cell_value = row[index]
+    if cell_value == '':
+      row[index] = None
+  for index in HAPROXY_STATS_INT_COLUMNS:
+    cell_value = row[index]
+    if cell_value is not MISSED:
+      row[index] = int(cell_value) if cell_value else None
 
 
 def get_stats_from_one_haproxy(socket_path, configs_dir):
   # Get CSV table with haproxy stats
   csv_buf = get_stats(socket_path)
   csv_buf.seek(2)  # Seek to the beginning but skip "# " in the first row
-  table = csv.DictReader(csv_buf, delimiter=',')
+  rows = iter(csv.reader(csv_buf, delimiter=','))
+  try:
+    header = next(rows)
+  except StopIteration:
+    # No data is in stats socket
+    return []
+
+  # Warn or fail if stats header doesn't contain expected columns
   if ProxiesStatsSource.first_run:
-    missed = ALL_HAPROXY_FIELDS - set(table.fieldnames)
-    if missed:
-      logger.warn("HAProxy stats fields {} are missed. Old version of HAProxy "
-                   "is probably used (v1.5+ is expected)".format(list(missed)))
+    if header[:len(HEADER_1_5_PLUS)] != HEADER_1_5_PLUS:
+      logger.warning("Old version of HAProxy is used (v1.5+ is expected).\n"
+                     "Actual header starts from:\n  {}\n"
+                     "Expected header should start from:\n  {}"
+                     .format(header, HEADER_1_5_PLUS))
+      if header[:len(ACCEPTABLE_HEADER)] != ACCEPTABLE_HEADER:
+        msg = ("HAProxy stats header doesn't contain expected columns.\n"
+               "Actual header starts from:\n  {}\n"
+               "Expected header should start from:\n  {}\n"
+               "Recommended HAProxy version is 1.5+."
+               .format(header, HEADER_1_5_PLUS))
+        logger.error(msg)
+        raise HAProxyStatsParsingError(msg)
     ProxiesStatsSource.first_run = False
 
   # Parse haproxy stats output line by line
   parsed_objects = defaultdict(list)
-  for row in table:
-    proxy_name = row['pxname']
+  for row in rows:
+    if not row or not row[0]:
+      # Skip last empty line
+      continue
+    _convert_ints_and_missing(row)
+    proxy_name = row[PXNAME]
     service = find_service_by_pxname(proxy_name)
-    svname = row['svname']
-    extra_values = {}
+    svname = row[SVNAME]
     if svname == 'FRONTEND':
-      stats_type = HAProxyFrontendStats
+      # User positional args to get performance improvement from slots
+      stats = HAProxyFrontendStats(
+        row[PXNAME],
+        row[SVNAME],
+        row[SCUR],
+        row[SMAX],
+        row[SLIM],
+        row[STOT],
+        row[BIN],
+        row[BOUT],
+        row[DREQ],
+        row[DRESP],
+        row[EREQ],
+        row[STATUS],
+        row[PID],
+        row[IID],
+        row[TYPE],
+        row[RATE],
+        row[RATE_LIM],
+        row[RATE_MAX],
+        row[HRSP_1XX],
+        row[HRSP_2XX],
+        row[HRSP_3XX],
+        row[HRSP_4XX],
+        row[HRSP_5XX],
+        row[HRSP_OTHER],
+        row[REQ_RATE],
+        row[REQ_RATE_MAX],
+        row[REQ_TOT],
+        row[COMP_IN],
+        row[COMP_OUT],
+        row[COMP_BYP],
+        row[COMP_RSP]
+      )
     elif svname == 'BACKEND':
-      stats_type = HAProxyBackendStats
-    elif row['qcur']:
+      # User positional args to get performance improvement from slots
+      stats = HAProxyBackendStats(
+        row[PXNAME],
+        row[SVNAME],
+        row[QCUR],
+        row[QMAX],
+        row[SCUR],
+        row[SMAX],
+        row[SLIM],
+        row[STOT],
+        row[BIN],
+        row[BOUT],
+        row[DREQ],
+        row[DRESP],
+        row[ECON],
+        row[ERESP],
+        row[WRETR],
+        row[WREDIS],
+        row[STATUS],
+        row[WEIGHT],
+        row[ACT],
+        row[BCK],
+        row[CHKDOWN],
+        row[LASTCHG],
+        row[DOWNTIME],
+        row[PID],
+        row[IID],
+        row[LBTOT],
+        row[TYPE],
+        row[RATE],
+        row[RATE_MAX],
+        row[HRSP_1XX],
+        row[HRSP_2XX],
+        row[HRSP_3XX],
+        row[HRSP_4XX],
+        row[HRSP_5XX],
+        row[HRSP_OTHER],
+        row[CLI_ABRT],
+        row[SRV_ABRT],
+        row[COMP_IN],
+        row[COMP_OUT],
+        row[COMP_BYP],
+        row[COMP_RSP],
+        row[LASTSESS],
+        row[QTIME],
+        row[CTIME],
+        row[RTIME],
+        row[TTIME]
+      )
+    elif row[QCUR] is not None:
       # Listener stats doesn't have "current queued requests" property
-      stats_type = HAProxyServerStats
       private_ip, port = service.get_ip_port_by_svname(svname)
-      extra_values['private_ip'] = private_ip
-      extra_values['port'] = port
+      # User positional args to get performance improvement from slots
+      stats = HAProxyServerStats(
+        private_ip,
+        port,
+        row[PXNAME],
+        row[SVNAME],
+        row[QCUR],
+        row[QMAX],
+        row[SCUR],
+        row[SMAX],
+        row[SLIM],
+        row[STOT],
+        row[BIN],
+        row[BOUT],
+        row[DRESP],
+        row[ECON],
+        row[ERESP],
+        row[WRETR],
+        row[WREDIS],
+        row[STATUS],
+        row[WEIGHT],
+        row[ACT],
+        row[BCK],
+        row[CHKFAIL],
+        row[CHKDOWN],
+        row[LASTCHG],
+        row[DOWNTIME],
+        row[QLIMIT],
+        row[PID],
+        row[IID],
+        row[SID],
+        row[THROTTLE],
+        row[LBTOT],
+        row[TRACKED],
+        row[TYPE],
+        row[RATE],
+        row[RATE_MAX],
+        row[CHECK_STATUS],
+        row[CHECK_CODE],
+        row[CHECK_DURATION],
+        row[HRSP_1XX],
+        row[HRSP_2XX],
+        row[HRSP_3XX],
+        row[HRSP_4XX],
+        row[HRSP_5XX],
+        row[HRSP_OTHER],
+        row[HANAFAIL],
+        row[CLI_ABRT],
+        row[SRV_ABRT],
+        row[LASTSESS],
+        row[LAST_CHK],
+        row[LAST_AGT],
+        row[QTIME],
+        row[CTIME],
+        row[RTIME],
+        row[TTIME]
+      )
     else:
-      stats_type = HAProxyListenerStats
-
-    stats_values = {
-      field: _get_field_value(row, field)
-      for field in attr.fields_dict(stats_type).keys()
-    }
-    stats_values.update(**extra_values)
-
-    stats = stats_type(**stats_values)
+      # User positional args to get performance improvement from slots
+      stats = HAProxyListenerStats(
+        row[PXNAME],
+        row[SVNAME],
+        row[SCUR],
+        row[SMAX],
+        row[SLIM],
+        row[STOT],
+        row[BIN],
+        row[BOUT],
+        row[DREQ],
+        row[DRESP],
+        row[EREQ],
+        row[STATUS],
+        row[PID],
+        row[IID],
+        row[SID],
+        row[TYPE],
+      )
     parsed_objects[proxy_name].append(stats)
+
+  net_connections = psutil.net_connections()
 
   # Attempt to merge separate stats object to ProxyStats instances
   proxy_stats_list = []
@@ -358,7 +617,7 @@ def get_stats_from_one_haproxy(socket_path, configs_dir):
       raise InvalidHAProxyStats(
         "Exactly one FRONTEND and one BACKEND line should correspond to "
         "a single proxy. Proxy '{}' has {} frontends and {} backends"
-          .format(proxy_name, len(frontends), len(backends))
+        .format(proxy_name, len(frontends), len(backends))
       )
 
     # Create ProxyStats object which contains all stats related to the proxy
@@ -366,7 +625,10 @@ def get_stats_from_one_haproxy(socket_path, configs_dir):
     application_id = service.get_application_id_by_pxname(proxy_name)
     try:
       bound_ip, bound_port = get_frontend_ip_port(configs_dir, proxy_name)
-      psutil_connections = get_connections(bound_ip, bound_port)
+      psutil_connections = sum(
+        1 for conn in net_connections
+        if conn.laddr == (bound_ip, bound_port) and conn.status == 'ESTABLISHED'
+      )
     except (OSError, IOError, BoundIpPortNotFound):
       psutil_connections = 0
     proxy_stats = ProxyStats(
@@ -415,7 +677,7 @@ class ProxiesStatsSource(object):
       proxies_stats=proxy_stats_list
     )
     logger.info("Prepared stats about {prox} proxies in {elapsed:.1f}s."
-                 .format(prox=len(proxy_stats_list), elapsed=time.time()-start))
+                .format(prox=len(proxy_stats_list), elapsed=time.time()-start))
     return stats