--- a/Hermes/appscale/hermes/producers/tests/test_proxy.py
+++ b/Hermes/appscale/hermes/producers/tests/test_proxy.py
@@ -9,7 +9,7 @@ from appscale.hermes.constants import MISSED
 from appscale.hermes.producers import proxy_stats
 
 CUR_DIR = os.path.dirname(os.path.realpath(__file__))
-TEST_DATA_DIR = os.path.join(CUR_DIR, 'test-data')
+DATA_DIR = os.path.join(CUR_DIR, 'test-data')
 
 
 class TestCurrentProxiesStats(unittest.TestCase):
@@ -22,9 +22,173 @@ class TestCurrentProxiesStats(unittest.TestCase):
       self.stats_file.close()
 
   @patch.object(proxy_stats.socket, 'socket')
+  def test_haproxy_stats_mapping(self, mock_socket):
+    # Mocking haproxy stats socket with csv file
+    # THE FILE CONTAINS MEANINGLESS DATA BUT IT HELP TO TEST COLUMNS MAPPING
+    self.stats_file = open(path.join(DATA_DIR, 'haproxy-stats-mapping.csv'))
+    fake_socket = MagicMock(recv=self.stats_file.read)
+    mock_socket.return_value = fake_socket
+
+    # Running method under test
+    stats_snapshot = proxy_stats.ProxiesStatsSource.get_current()
+    proxy_a = stats_snapshot.proxies_stats[0]
+
+    # Verify Frontend stats
+    self.assertEqual(
+      proxy_a.frontend,
+      proxy_stats.HAProxyFrontendStats(
+        pxname='ProxyA',
+        svname='FRONTEND',
+        scur=0,
+        smax=1,
+        slim=2,
+        stot=3,
+        bin=4,
+        bout=5,
+        dreq=6,
+        dresp=7,
+        ereq=8,
+        status='OPEN',
+        pid=9,
+        iid=10,
+        type=12,
+        rate=13,
+        rate_lim=14,
+        rate_max=15,
+        hrsp_1xx=16,
+        hrsp_2xx=17,
+        hrsp_3xx=18,
+        hrsp_4xx=19,
+        hrsp_5xx=20,
+        hrsp_other=21,
+        req_rate=22,
+        req_rate_max=23,
+        req_tot=24,
+        comp_in=25,
+        comp_out=26,
+        comp_byp=27,
+        comp_rsp=28,
+      )
+    )
+
+    # Verify Backend stats
+    self.assertEqual(
+      proxy_a.backend,
+      proxy_stats.HAProxyBackendStats(
+        pxname='ProxyA',
+        svname='BACKEND',
+        qcur=200,
+        qmax=201,
+        scur=202,
+        smax=203,
+        slim=204,
+        stot=205,
+        bin=206,
+        bout=207,
+        dreq=208,
+        dresp=209,
+        econ=210,
+        eresp=211,
+        wretr=212,
+        wredis=213,
+        status='UP',
+        weight=214,
+        act=215,
+        bck=216,
+        chkdown=217,
+        lastchg=218,
+        downtime=219,
+        pid=220,
+        iid=221,
+        lbtot=223,
+        type=224,
+        rate=225,
+        rate_max=226,
+        hrsp_1xx=227,
+        hrsp_2xx=228,
+        hrsp_3xx=229,
+        hrsp_4xx=230,
+        hrsp_5xx=231,
+        hrsp_other=232,
+        cli_abrt=233,
+        srv_abrt=234,
+        comp_in=235,
+        comp_out=236,
+        comp_byp=237,
+        comp_rsp=238,
+        lastsess=239,
+        qtime=240,
+        ctime=241,
+        rtime=242,
+        ttime=243,
+      )
+    )
+
+    # Verify Server stats
+    self.assertEqual(
+      proxy_a.servers[0],
+      proxy_stats.HAProxyServerStats(
+        private_ip=None,
+        port=None,
+        pxname='ProxyA',
+        svname='ProxyA-10.10.8.28:17447',
+        qcur=100,
+        qmax=101,
+        scur=102,
+        smax=103,
+        slim=104,
+        stot=105,
+        bin=106,
+        bout=107,
+        dresp=108,
+        econ=109,
+        eresp=110,
+        wretr=111,
+        wredis=112,
+        status='UP',
+        weight=113,
+        act=114,
+        bck=115,
+        chkfail=116,
+        chkdown=117,
+        lastchg=118,
+        downtime=119,
+        qlimit=None,
+        pid=120,
+        iid=121,
+        sid=122,
+        throttle=None,
+        lbtot=123,
+        tracked=None,
+        type=124,
+        rate=125,
+        rate_max=126,
+        check_status='L4OK',
+        check_code=None,
+        check_duration=127,
+        hrsp_1xx=128,
+        hrsp_2xx=129,
+        hrsp_3xx=130,
+        hrsp_4xx=131,
+        hrsp_5xx=132,
+        hrsp_other=133,
+        hanafail=134,
+        cli_abrt=135,
+        srv_abrt=136,
+        lastsess=137,
+        last_chk='138',
+        last_agt='139',
+        qtime=140,
+        ctime=141,
+        rtime=142,
+        ttime=143,
+      )
+    )
+
+  @patch.object(proxy_stats.socket, 'socket')
   def test_haproxy_stats_v1_5(self, mock_socket):
     # Mocking haproxy stats socket with csv file
-    self.stats_file = open(path.join(TEST_DATA_DIR, 'haproxy-stats-v1.5.csv'))
+    self.stats_file = open(path.join(DATA_DIR, 'haproxy-stats-v1.5.csv'))
     fake_socket = MagicMock(recv=self.stats_file.read)
     mock_socket.return_value = fake_socket
 
@@ -74,23 +238,24 @@ class TestCurrentProxiesStats(unittest.TestCase):
     self.assertEqual(dashboard.listeners, [])
 
   @patch.object(proxy_stats.socket, 'socket')
-  @patch.object(proxy_stats.logger, 'warn')
+  @patch.object(proxy_stats.logger, 'warning')
   def test_haproxy_stats_v1_4(self, mock_logging_warn, mock_socket):
     # Mocking "echo 'show stat' | socat stdio unix-connect:{}" with csv file
-    self.stats_file = open(path.join(TEST_DATA_DIR, 'haproxy-stats-v1.4.csv'))
+    self.stats_file = open(path.join(DATA_DIR, 'haproxy-stats-v1.4.csv'))
     fake_socket = MagicMock(recv=self.stats_file.read)
     mock_socket.return_value = fake_socket
 
     # Running method under test
+    proxy_stats.ProxiesStatsSource.first_run = True
     stats_snapshot = proxy_stats.ProxiesStatsSource.get_current()
 
     # Verifying outcomes
     self.assertIsInstance(stats_snapshot.utc_timestamp, float)
     proxies_stats = stats_snapshot.proxies_stats
-    mock_logging_warn.assert_called_once_with(
-      "HAProxy stats fields ['rtime', 'ctime', 'comp_in', 'qtime', 'comp_byp', "
-      "'lastsess', 'comp_rsp', 'last_chk', 'ttime', 'comp_out', 'last_agt'] "
-      "are missed. Old version of HAProxy is probably used (v1.5+ is expected)"
+    self.assertTrue(
+      mock_logging_warn.call_args[0][0].startswith(
+        "Old version of HAProxy is used (v1.5+ is expected)."
+      )
     )
     self.assertEqual(len(proxies_stats), 5)
     proxies_stats_dict = {
@@ -145,7 +310,7 @@ class TestCurrentProxiesStats(unittest.TestCase):
 
 class TestGetServiceInstances(unittest.TestCase):
   def setUp(self):
-    stats_file = open(path.join(TEST_DATA_DIR, 'haproxy-stats-v1.5.csv'))
+    stats_file = open(path.join(DATA_DIR, 'haproxy-stats-v1.5.csv'))
     fake_socket = MagicMock(recv=stats_file.read)
     self.socket_patcher = patch.object(proxy_stats.socket, 'socket')
     socket_mock = self.socket_patcher.start()
@@ -188,3 +353,6 @@ class TestGetServiceInstances(unittest.TestCase):
     unknown = proxy_stats.get_service_instances('mocked', 'gae_not_running')
     self.assertEqual(unknown, [])
 
+
+if __name__ == '__main__':
+  unittest.main()