… Hermes/appscale/hermes/producers/tests/test-data/haproxy-stats-mapping.csv