This PR does not change any API or functionality.

It fixes 4 performance issues:
 - AppController should be fine to get cached version of proxy stats with age not older than 5 seconds.
 - Hermes should call `psutil.net_connections()` only once per `get_current()` call.
 - Use absolute integer indexes when parsing CSV output from haproxy socket (instead of `csv.DictReader`).
 - Use explicit attribute assignment instead of keyword argument dict.

Performance improvement should be noticed only on Load Balancer nodes.

### Demo
CPU times were measured for Hermes process with 30 minutes interval before and after this PR.
Deployment was in idle state and had guestbook, hawkeye and dashboard deployed.
The only requests Hermes was serving were from deployment itself with no profiling.

#### Master version
CPU User: 64.8, System: 47.1
Which is ~3.7 CPU seconds per minute.

#### Branch
CPU User: 12.8, System: 2.7
Which is ~0.5 CPU seconds per minute.