@jeanleonov 2019-04-10
## Demo
Here is a fragment of hermes logs corresponding to the moment when AppContrller analyses applications load. AC loops though all services deployed and does hermes call for each of them (it could make it once instead, but I didn't want to do too many changes in AC). Anyway, proper use of Hermes caching + Hermes side performance optimization are good to have and resolve the problem.

Before:
```
2019-04-10 13:15:55,737 DEBUG proxy_stats.py:409 Processing services haproxy stats 
2019-04-10 13:15:55,800 DEBUG proxy_stats.py:409 Processing apps haproxy stats 
2019-04-10 13:15:55,800 INFO proxy_stats.py:418 Prepared stats about 4 proxies in 0.1s. 
2019-04-10 13:15:55,804 INFO web.py:1908 200 GET /stats/local/proxies (10.0.2.15) 68.10ms 
2019-04-10 13:15:56,681 DEBUG proxy_stats.py:409 Processing services haproxy stats 
2019-04-10 13:15:56,734 DEBUG proxy_stats.py:409 Processing apps haproxy stats 
2019-04-10 13:15:56,735 INFO proxy_stats.py:418 Prepared stats about 5 proxies in 0.1s. 
2019-04-10 13:15:56,740 INFO web.py:1908 200 GET /stats/local/proxies (10.0.2.15) 59.07ms 
2019-04-10 13:15:56,752 DEBUG proxy_stats.py:409 Processing services haproxy stats 
2019-04-10 13:15:56,797 DEBUG proxy_stats.py:409 Processing apps haproxy stats 
2019-04-10 13:15:56,798 INFO proxy_stats.py:418 Prepared stats about 5 proxies in 0.0s. 
2019-04-10 13:15:56,803 INFO web.py:1908 200 GET /stats/local/proxies (10.0.2.15) 51.61ms 
```
Caching is not used and every call takes about 60ms.

After:
```
2019-04-10 13:26:27,623 DEBUG proxy_stats.py:759 Processing services haproxy stats 
2019-04-10 13:26:27,626 DEBUG proxy_stats.py:759 Processing apps haproxy stats 
2019-04-10 13:26:27,626 INFO proxy_stats.py:769 Prepared stats about 5 proxies in 0.0s. 
2019-04-10 13:26:27,635 INFO web.py:1908 200 GET /stats/local/proxies (10.0.2.15) 23.01ms 
2019-04-10 13:26:27,657 INFO handlers.py:79 Returning cached snapshot with age 0.66s 
2019-04-10 13:26:27,663 INFO web.py:1908 200 GET /stats/local/proxies (10.0.2.15) 6.10ms 
2019-04-10 13:26:32,389 INFO handlers.py:79 Returning cached snapshot with age 5.39s 
2019-04-10 13:26:32,391 INFO web.py:1908 200 GET /stats/local/proxies (10.0.2.15) 3.53ms 
```
Caching is used, the first call with actual work (which was optimized) takes 23ms, the rest of them take 4-6ms.